v2.4.13: 修复构建警告和lint错误，优化批量出售功能
- 修复AutoboxingStateCreation lint警告: mutableStateOf(1)改为mutableIntStateOf(1)
- 修复SuspiciousIndentation lint错误: ItemDetailDialog属性展示if语句添加大括号
- 批量出售对话框新增确认弹窗，显示物品数量和获得灵石
- 出售价格计算统一使用GameConfig.Rarity.calculateSellPrice方法
- 修复物品详情对话框缺少暴击率/暴击效果/丹药类别显示
- 统一百分比格式化和丹药类别标签

+++ android/app/src/main/java/com/xianxia/sect/ui/components/ItemCard.kt
@@ -28,7 +28,7 @@ data class ItemCardData(
     val type: String? = null,
     val stats: Map<String, Int> = emptyMap(),
     val additionalInfo: String? = null,
-    val price: Int = 0,
+    val price: Long = 0L,
     val grade: String? = null,
     val isLocked: Boolean = false
 )
